# assets/docs/rider/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "### 说明\n接口采用非标准Restful API，所有http返回代码均为`200`，当返回为非`200`时应为network错误，需要及时排查。\n\u003cbr\u003e\n接口返回说明查看 **[返回](#返回)**\n\u003cbr\u003e\n图片/附件前缀 `https://cdn.auroraride.com/`\n\n\u003cbr /\u003e\n\n\u003cbr /\u003e\n\n### 接口编号\n\n第一位代表接口端分类\n\n- M 管理端\n- R 骑手端\n- E 门店端\n- C 通用\n\n第二位代表子分类（36进制）\n\n后三位代表子编号（10进制）\n\n\u003cbr /\u003e\n\n\u003cbr /\u003e\n\n\n### 认证\n项目接口使用简单认证，认证方式为`header`中添加对应的认证`token`\n|  header   |  类型  |  接口  |\n| :-----: | :----: | :--: |\n|  X-Rider-Token   |  string   |  骑手API  |\n| X-Manager-Token | string |  后台API  |\n|  X-Employee-Token   | string |  员工API  |\n\n\u003cbr /\u003e\n\n\u003cbr /\u003e\n\n### 返回\n\n一个标准的返回应包含以下结构\n\n|  字段   |  类型  |  必填  |  说明  |\n| :-----: | :----: | :--: | :--: |\n|  code   |  int   |  是  |  返回代码  |\n| message | string |  是  |  返回消息  |\n|  data   | object |  是  |  返回数据  |\n\n`code`代码取值说明\n\n| 十进制 | 十六进制 | 说明 |\n| :----: | :------: | :--: |\n| 0  |  0x000  | 请求成功 |\n| 256 |  0x100  | 请求失败 |\n| 512 |  0x200  | *需要认证(跳转登录) |\n| 768 |  0x300  | *用户被封禁 |\n| 1024 |  0x400  | 资源未获 |\n| 1280 |  0x500  | 未知错误 |\n| 1536 |  0x600  | *需要实名 |\n| 1792 |  0x700  | *需要验证 (更换设备, 需要人脸验证) |\n| 2048 |  0x800  | *需要联系人 |\n| 2304 |  0x900  | 请求过期 |\n\n\u003e 当返回值是`1792(0x700)需要人脸验证`或`1536(0x600)需要实名`的时候`data`返回值为`{\"url\": \"string\"}`, 直接跳转url\n\n\n比如：\n\u003e 默认成功返回\n```json\n{\n  \"code\": 0,\n  \"message\": \"OK\",\n  \"data\": {\n    \"status\": true\n  }\n}\n```\n",
...
    },
    "basePath": "/",
    "paths": {
...
        "/rider/v1/enterprise/battery": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enterprise - 团签"
                ],
                "summary": "R3010 企业骑手获取可用电池",
                "operationId": "RiderEnterpriseBattery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "城市ID",
                        "name": "cityId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rider/v1/enterprise/exit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enterprise - 团签"
                ],
                "summary": "R3017 退出团签",
                "operationId": "RiderEnterpriseExitEnterprise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    }
                }
            }
        },
        "/rider/v1/enterprise/info": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enterprise - 团签"
                ],
                "summary": "R3016 骑手团签信息",
                "operationId": "RiderEnterpriseRiderEnterpriseInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "团签id",
                        "name": "enterpriseId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "站点id",
                        "name": "stationId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.EnterproseInfoRsp"
                        }
                    }
                }
            }
        },
        "/rider/v1/enterprise/join": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enterprise - 团签"
                ],
                "summary": "R3015 企业骑手加入团签",
                "operationId": "RiderEnterpriseJoinEnterprise",
                "parameters": [
                    {
                        "description": "加入团签请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EnterproseInfoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/rider/v1/enterprise/subscribe": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enterprise - 团签"
                ],
                "summary": "R3012 企业骑手订阅激活状态",
                "operationId": "RiderEnterpriseSubscribeStatus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "订阅ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TRUE已激活, FALSE未激活",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enterprise - 团签"
                ],
                "summary": "R3011 企业骑手选择电池",
                "operationId": "RiderEnterpriseSubscribe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "电池选择请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EnterpriseRiderSubscribeChooseReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.EnterpriseRiderSubscribeChooseRes"
                        }
                    }
                }
            }
        },
        "/rider/v1/enterprise/subscribe/alter": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enterprise - 团签"
                ],
                "summary": "R3013 加时申请",
                "operationId": "RiderEnterpriseSubscribeAlter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "申请增加订阅时长请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SubscribeAlterRiderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    }
                }
            }
        },
        "/rider/v1/enterprise/subscribe/alter/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enterprise - 团签"
                ],
                "summary": "R3014 加时申请列表",
                "operationId": "RiderEnterpriseSubscribeAlterList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前页, 从1开始, 默认1",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数据, 默认20",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "审核状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.PaginationRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.SubscribeAlterApplyListRes"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.City": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "城市ID",
                    "type": "integer"
                },
                "name": {
                    "description": "城市",
                    "type": "string"
                }
            }
        },
...
        "model.EbikeBrand": {
            "type": "object",
            "properties": {
                "cover": {
                    "description": "封面图",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                }
            }
        },
...
        "model.EnterprisePriceWithCity": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "城市",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.City"
                        }
                    ]
                },
                "ebikeBrand": {
                    "description": "车辆型号",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.EbikeBrand"
                        }
                    ]
                },
                "id": {
                    "description": "价格ID",
                    "type": "integer"
                },
                "intelligent": {
                    "description": "是否智能电池",
                    "type": "boolean"
                },
                "model": {
                    "description": "电池型号",
                    "type": "string"
                },
                "price": {
                    "description": "单价(元/天)",
                    "type": "number"
                }
            }
        },
        "model.EnterpriseRiderSubscribeChooseReq": {
            "type": "object",
            "required": [
                "model"
            ],
            "properties": {
                "model": {
                    "type": "string"
                }
            }
        },
        "model.EnterpriseRiderSubscribeChooseRes": {
            "type": "object",
            "properties": {
                "qrcode": {
                    "description": "二维码, 格式为SUBSCRIBE:订阅ID, 后续使用订阅ID请求状态",
                    "type": "string"
                }
            }
        },
...
        "model.EnterproseInfoReq": {
            "type": "object",
            "required": [
                "enterpriseId",
                "stationId"
            ],
            "properties": {
                "enterpriseId": {
                    "description": "团签id",
                    "type": "integer"
                },
                "stationId": {
                    "description": "站点id",
                    "type": "integer"
                }
            }
        },
        "model.EnterproseInfoRsp": {
            "type": "object",
            "properties": {
                "days": {
                    "description": "天数",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "enterproseName": {
                    "description": "团签名称",
                    "type": "string"
                },
                "isJoin": {
                    "description": "是否可以加入团签",
                    "type": "boolean"
                },
                "priceList": {
                    "description": "价格",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.EnterprisePriceWithCity"
                    }
                },
                "stationName": {
                    "description": "站点名称",
                    "type": "string"
                }
            }
        },
...
        "model.Pagination": {
            "type": "object",
            "properties": {
                "current": {
                    "description": "当前页",
                    "type": "integer"
                },
                "pages": {
                    "description": "总页数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
        "model.PaginationRes": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "返回数据"
                },
                "pagination": {
                    "description": "分页属性",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Pagination"
                        }
                    ]
                }
            }
        },
...
        "model.Rider": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                },
                "phone": {
                    "description": "电话",
                    "type": "string"
                }
            }
        },
...
        "model.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "model.SubscribeAlterApplyListRes": {
            "type": "object",
            "properties": {
                "applyTime": {
                    "description": "申请时间",
                    "type": "string"
                },
                "days": {
                    "description": "天数",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "operator": {
                    "description": "操作人",
                    "type": "string"
                },
                "reviewTime": {
                    "description": "审核时间",
                    "type": "string"
                },
                "rider": {
                    "description": "骑手",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Rider"
                        }
                    ]
                },
                "status": {
                    "description": "审核状态",
                    "type": "integer"
                },
                "subscribeEndAt": {
                    "description": "团签到期时间",
                    "type": "string"
                }
            }
        },
        "model.SubscribeAlterRiderReq": {
            "type": "object",
            "required": [
                "days"
            ],
            "properties": {
                "days": {
                    "type": "integer"
                }
            }
        },
...
    }
}